# event-ticketing-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Ticket API",
        "description": "API for event ticket purchasing system",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get paginated list of events",
                "operationId": "getEvents",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Events retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create event (Admin only)",
                "operationId": "createEvent",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "location",
                                    "event_date",
                                    "sale_start_date",
                                    "sale_end_date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Summer Music Festival"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Annual summer music festival"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Central Park, New York"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "event_date": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2025-07-15 18:00:00"
                                    },
                                    "sale_start_date": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2025-06-01 10:00:00"
                                    },
                                    "sale_end_date": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2025-07-14 23:59:59"
                                    },
                                    "allow_cancellation": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "cancellation_hours_before": {
                                        "type": "integer",
                                        "example": 24
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get single event with ticket types",
                "operationId": "getEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event retrieved successfully"
                    }
                }
            }
        },
        "/api/events/{id}/ticket-types": {
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Add ticket type to event (Admin only)",
                "operationId": "addTicketType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "total_quantity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "VIP"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "VIP access with premium amenities"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 199.99
                                    },
                                    "total_quantity": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket type added successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}